requirements.txt
altair==4.2.2
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.2
extra-streamlit-components==0.1.80
fonttools==4.57.0
fpdf==1.7.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.0
mdurl==0.1.2
narwhals==1.37.1
numpy==1.26.4
openpyxl==3.1.0
packaging==23.2
pandas==2.2.0
pillow==10.4.0
pluggy==1.5.0
protobuf==4.25.7
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.12.2
six==1.17.0
smmap==5.0.2
streamlit==1.32.2
streamlit-authenticator==0.2.3
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
